+++ src/main/java/com/example/sns/alarm/presentation/dto/AlarmDto.java
@@ -1,4 +1,4 @@
-package com.example.sns.alarm.ui.dto;
+package com.example.sns.alarm.presentation.dto;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
